pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "a-test-automation"
version = "0.1.0"
description = "Simple, extendable test automation framework with pytest, Playwright, and requests"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "allure-pytest==2.13.5",
  "playwright==1.50.0",
  "pytest==8.3.4",
  "pytest-asyncio==0.25.3",
  "python-dotenv==1.0.1",
  "requests==2.32.3",
  "deepeval>=1.2.0,<2.0.0",
  "pytest-xdist==3.6.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
